Move some classes between package
