+++ src/main/java/ldbc/snb/datagen/serializer/snb/csv/mergeforeign/CSVMergeForeignDynamicPersonSerializer.java
@@ -40,7 +40,7 @@ import ldbc.snb.datagen.objects.Knows;
 import ldbc.snb.datagen.objects.Person;
 import ldbc.snb.datagen.objects.StudyAt;
 import ldbc.snb.datagen.objects.WorkAt;
-import ldbc.snb.datagen.serializer.HDFSCSVWriter;
+import ldbc.snb.datagen.writer.HDFSCSVWriter;
 import ldbc.snb.datagen.serializer.DynamicPersonSerializer;
 import org.apache.hadoop.conf.Configuration;
