+++ src/main/java/ldbc/snb/datagen/serializer/snb/turtle/TurtleDynamicActivitySerializer.java
@@ -37,9 +37,8 @@ package ldbc.snb.datagen.serializer.snb.turtle;
 
 import ldbc.snb.datagen.dictionary.Dictionaries;
 import ldbc.snb.datagen.objects.*;
-import ldbc.snb.datagen.serializer.HDFSWriter;
+import ldbc.snb.datagen.writer.HDFSWriter;
 import ldbc.snb.datagen.serializer.DynamicActivitySerializer;
-import ldbc.snb.datagen.serializer.Turtle;
 import ldbc.snb.datagen.vocabulary.*;
 import org.apache.hadoop.conf.Configuration;
