+++ src/main/java/ldbc/snb/datagen/serializer/snb/turtle/TurtleStaticSerializer.java
@@ -40,9 +40,8 @@ import ldbc.snb.datagen.objects.Organization;
 import ldbc.snb.datagen.objects.Place;
 import ldbc.snb.datagen.objects.Tag;
 import ldbc.snb.datagen.objects.TagClass;
-import ldbc.snb.datagen.serializer.HDFSWriter;
+import ldbc.snb.datagen.writer.HDFSWriter;
 import ldbc.snb.datagen.serializer.StaticSerializer;
-import ldbc.snb.datagen.serializer.Turtle;
 import ldbc.snb.datagen.vocabulary.*;
 import org.apache.hadoop.conf.Configuration;
